% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{add_error}
\alias{add_error}
\title{Add error to a specific columns in a table}
\usage{
add_error(df, vars, factor = 1)
}
\arguments{
\item{df}{a data frame}

\item{vars}{the names or indices of columns}

\item{factor}{numeric for factor argument in jitter function}
}
\description{
Add error to a specific columns in a table
}

