% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templates.R
\name{add_gauss_error}
\alias{add_gauss_error}
\title{Add Gaussian error}
\usage{
add_gauss_error(.data, vars, noise)
}
\arguments{
\item{.data}{a data frame}

\item{vars}{names or indices of the data.frame to add the error.}

\item{noise}{the desired amount of standard error in the noise to add to the variables.
Note that after noise is added the var is rescaled to 0 and 1.}
}
\description{
In the creation of a data frame from a simulated system,
this adds Gaussian error to the variables of a table and rescales them
to between 0 and 1.
}

