% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structure.R
\name{addBiases}
\alias{addBiases}
\title{Add Nodes Corresponding to Biases}
\usage{
addBiases(g)
}
\arguments{
\item{g}{igraph object}
}
\value{
an igraph object
}
\description{
If biases already exist in the graph, no biases are added.
This function uses the igraph method on the `+` generic. In this code it is called
explicitly with igraph::`+.igraph` for safety reasons.
}

