% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structure.R
\name{addDataToVertices}
\alias{addDataToVertices}
\title{Add data to vertices}
\usage{
addDataToVertices(g, data)
}
\arguments{
\item{g}{igraph object. The vertices must be named.}

\item{data}{a data frame. All of the names in the data from must match a vertex name.}
}
\description{
Adds data to Any vertex that does not map to a name in the data will be treated as a hidden node.
If a variable in the data is not present amongst the graph vertices, an error is thrown.
}

