% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structure.R
\name{addInterceptNodes}
\alias{addInterceptNodes}
\title{Add Nodes Corresponding to Biases}
\usage{
addInterceptNodes(g)
}
\description{
If intercepts (biases) already exist in the graph, no intercepts are added.
}

