% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{calculateVals}
\alias{calculateVals}
\title{Get deviance}
\usage{
calculateVals(g, v)
}
\arguments{
\item{g,}{the graph model}

\item{v.index,}{the index of a vertex}
}
\value{
an updated graph model
}
\description{
Deviance is defined as residual sum squares on all the output variables and their predictions
}
\details{
Calculate the signal propagation in the signalgraph model

For a given vertex, the values for the vertex attribute 'input.signal' are calculated as the linear
combination of outputs with edge weights as weights. 'output.signal' is calculated by applying the activation
function to the input signal.
}

