% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templates.R
\name{case_from_graph}
\alias{case_from_graph}
\title{Test Case: Random graph and matching data}
\usage{
case_from_graph(g, n)
}
\arguments{
\item{n}{the number of desired rows in the data}

\item{m}{the number of desired nodes}
}
\value{
A list of two elements, graph and data.
}
\description{
Generates an igraph object and a corresponding dataset.  A number of unobserved vertices will
be chosen at random.  Used to randomly generate test cases.
}

