% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal-prioritization.R
\name{causal_prioritization}
\alias{causal_prioritization}
\title{K order node selection}
\usage{
causal_prioritization(g, k)
}
\description{
Ranks vertices in the signal graph by a given ranking method.  Chooses the K
vertices with the top K highest ranking.  Returns the union of the causal 
neighborhoods of those K vertices.
}

