% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structure.R
\name{checkArgs}
\alias{checkArgs}
\title{Check validity of arguments}
\usage{
checkArgs(g, data)
}
\arguments{
\item{g}{an igraph object, vertices should be named.}

\item{data}{a dataframe, each variable name should match a vertex name in the graph}
}
\value{
the input graph, ready for passing to a subsequent function
}
\description{
Check validity of both inputs to constuction of signalgraph object
}

