% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structure.R
\name{checkVector}
\alias{checkVector}
\title{Checks the validity of a vertex vector attribute}
\usage{
checkVector(item)
}
\description{
Checks the validity of a vertex vector attribute
}
\examples{
g <- get_gate()
checkVector(unlist(V(g)[is.observed]$observed))
}

