% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templates.R
\name{donate_edges}
\alias{donate_edges}
\title{Donate edges to degree-less vertices}
\usage{
donate_edges(g)
}
\arguments{
\item{g}{an igraph output of lucy::power_law_sim}
}
\value{
the graph rewired such that there are no degreeless vertices
}
\description{
Used to add edges to nodes that lack edges in graphs generated by power_signal_graph.
A 'rich get poorer' approach is used, wherein nodes with many incoming edges 'donate'
an incoming edge to an edgeless node.
}

