% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{examine_signal_graph}
\alias{examine_signal_graph}
\title{Summarize the values of an signal graph model}
\usage{
examine_signal_graph(g)
}
\arguments{
\item{g}{a signal graph object}
}
\description{
A summary function that prints out the values of various graph, edge, and vertex attributes in formatted lists.
}

