% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit_initialized_sg}
\alias{fit_initialized_sg}
\title{Fit an initialized network}
\usage{
fit_initialized_sg(g, min.iter = 2, max.iter = 5, epsilon = 1e-04,
  verbose = FALSE)
}
\arguments{
\item{g}{an igraph object initialized to an unfit signalgraph object}

\item{min.iter}{minimum number of iterations}

\item{max.iter}{maximum number of iterations}

\item{epsilon}{after the minimum number of iterations, when change in means squared error
between iterations falls below epsilon, cease optimizing and return current estimates}

\item{verbose}{if TRUE prints info on propagation}
}
\description{
Fit an initialized network
}

