% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit_weights_for_node}
\alias{fit_weights_for_node}
\title{Fit incoming edge weights for intervention vertex}
\usage{
fit_weights_for_node(g, v)
}
\arguments{
\item{g}{a signal graph object}

\item{v}{the vertex whose incoming edge weights are to be optimized}
}
\description{
Constructs a loss function specific to the vertex, and uses this to optimize the weights on the
vertices incoming edges using the current weights as the initial values in the optimization.
The incoming edge weights are replaced with the optimized values, and the updated graph object.
is returned.
}
\seealso{
\code{\link{fit_weights_for_node}}, \code{\link{getObjective}}, \code{\link{getOptimizationFunction}}, \code{\link{getOptimizationFunctionNG}}, \code{\link{getPrediction}}
}

