% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{fitInitializedNetwork}
\alias{fitInitializedNetwork}
\title{Fit an initialized network}
\usage{
fitInitializedNetwork(g, epsilon = 0.001, max.iter = 100, verbose = F)
}
\description{
g an initialized graph
epsilon when means square area falls below epsilon, stop
max.iter maximum number of iterations
verbose if TRUE print messages generated during optimization
}

