% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimization.R
\name{fitWeightsForEdgeTarget}
\alias{fitWeightsForEdgeTarget}
\title{Fit edge weight}
\usage{
fitWeightsForEdgeTarget(g, e)
}
\arguments{
\item{g}{a signalgraph object}

\item{e}{an index for the edge in the signalgraph object}
}
\value{
an updated signalgraph object
}
\description{
Adjust the weight for a given edge by minimizing a loss function.  This function takes a signal edge
as its argument.  It identifies the target vertex of the edge, then calls fitWeightsForNode,
which optimizes the loss function over all incoming edges to that target vertex, rather than just the
edge argument alone.  Once the weight has been optimized, the updated attribute is set to true.
}
\seealso{
fitWeightsForNode
}

