% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{format_vertex_list}
\alias{format_vertex_list}
\title{A helper function used in summarizing signal graph structures}
\usage{
format_vertex_list(output)
}
\description{
A helper function used in summarizing signal graph structures
}
\seealso{
examine_signal_graph
}

