% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{get_deviance}
\alias{get_deviance}
\title{Get deviance}
\usage{
get_deviance(g)
}
\arguments{
\item{g,}{a fitted signal graph model}
}
\description{
Deviance is defined as residual sum squares on all the output variables and their predictions
}

