% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{get_downstream_vertices}
\alias{get_downstream_vertices}
\title{Given a vertex, find vertices with edges that need optimizing first}
\usage{
get_downstream_vertices(g, v)
}
\arguments{
\item{g}{a signalgraph object}

\item{v}{an vertex index in the signal graph}
}
\value{
set of vertex indices
}
\description{
This algorithm takes all the incoming edge weights for a given vertex and optimizes them at
once.  When optimizing weights, before optimizing the objective function for a
given weight, it is necessary to first optimize for all weights that are no paths downstream
from the given weight and upstream of an observed variable.  This function identifies all such
downstream vertices that posses incoming weights that edges that, relative to any
of the given vertices incoming edges, need to be optimized first.
}

