% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{get_fitted}
\alias{get_fitted}
\title{Pull fitted values from a signal graph into a data frame}
\usage{
get_fitted(g)
}
\arguments{
\item{g}{A fitted graph}
}
\value{
a data frame containing fitted values
}
\description{
Pull fitted values from a signal graph into a data frame
}
\seealso{
recover_design
}

