% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templates.R
\name{get_gate}
\alias{get_gate}
\title{Create an signal graph model of a multi-layer perceptron logic gate (unfit case).}
\usage{
get_gate(outputs = "all", layers = NULL)
}
\arguments{
\item{outputs}{character vector describing the desired gates.  Defaults to "all", in which case
(AND, OR, NAND, NOR, XOR, and XNOR) are the outputs.  Otherwise any subset of these will work.}

\item{layers}{a numeric vector where each element corresponds to a layer, and the value of an element
is the number of hidden nodes in the layer.  Default is NULL.  If value is null, no hidden layers will be
present (just an input and output layer).}
}
\value{
a signal graph model object with unfit weights
}
\description{
Creates an signal graph model of a multi-layer perceptron logic gate.  The signal graph weights are not fit.
This is used primarily for experiments.
}

