% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{get_structure}
\alias{get_structure}
\title{Pull igraph structure from a signal graph}
\usage{
get_structure(g)
}
\arguments{
\item{g}{a signalgraph object}
}
\description{
Pull igraph structure from a signal graph
}

