% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{get_vertex_mse}
\alias{get_vertex_mse}
\title{Get the mean squared error for a given vertex}
\usage{
get_vertex_mse(g, v)
}
\arguments{
\item{g}{a signalgraph object}

\item{v}{a vertex id in the signal graph}
}
\value{
means squared error calculation
}
\description{
Get the mean squared error for a given vertex
}

