% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{getDependentEdges}
\alias{getDependentEdges}
\title{Determine the edges in g whose weights impact the optimization of the weight of edge e}
\usage{
getDependentEdges(g, e)
}
\description{
Determine the edges in g whose weights impact the optimization of the weight of edge e
}

