% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backpropagation.R
\name{getGradientFunction}
\alias{getGradientFunction}
\title{Gradient generating closure}
\usage{
getGradientFunction(g, v)
}
\description{
In the network, any given node v has a set of incoming edges.  Those incoming edges have weights.
The fitting procedure fits the weights in the network by seperately these weights for each node.
More specifically, a seperate optimization is conducted for each node v, where in the weights on the
incoming edges to v are optimized on the loss function.
}
\details{
This closure returns a function that computes the gradient of all the weights
on the incoming edges of a node v.  The output function is intended to be passed to an optimization
functional.  I use a closure because the optim functional only accepts a weights are
}

