% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimization.R
\name{getLossFunction}
\alias{getLossFunction}
\title{Create a loss function dependent on a given vertex}
\usage{
getLossFunction(g, v)
}
\description{
This is a closure that creates a loss function that varies given a weight vector.
The weights are weight attributes of incoming edges to the vertex given in the argument.
Thus for a given vertex, you can inspect how varying the weights that determine the vertex affect loss.
}

