% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{getMSE}
\alias{getMSE}
\alias{vertexMSE}
\alias{vertexMSEs}
\title{MSE calculation}
\usage{
getMSE(g)

vertexMSE(g, v)

vertexMSEs(g)
}
\arguments{
\item{g}{a signal graph object}

\item{v}{a vertex id in the signal graph}
}
\description{
vertexMSE calculates the mean squared error for a given fitted variables in a signalgraph
object. vertexMSEs calculates MSE for each fitted vertices and returns an array of values.
getMSE calculates the mean squared error for all the fitted variables in a signalgraph
object, then reports the mean of means.
}

