% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backpropagation.R
\name{getOptimizationFunction}
\alias{getOptimizationFunction}
\title{Closure for optimization with BFGS and specified gradient function}
\usage{
getOptimizationFunction(g, lossFunction, getGradient)
}
\description{
Closure for optimization with BFGS and specified gradient function
}

