% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{getOptimizationFunctionNG}
\alias{getOptimizationFunctionNG}
\title{Closure for optimization with BFGS, no gradient function specified}
\usage{
getOptimizationFunctionNG(g, lossFunction)
}
\description{
Closure for optimization with BFGS, no gradient function specified
}
\seealso{
\code{\link{fit_weights_for_node}}, \code{\link{getObjective}}, \code{\link{getOptimizationFunction}}, \code{\link{getOptimizationFunctionNG}}, \code{\link{getPrediction}}
}

