% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimization.R
\name{getPrediction}
\alias{getPrediction}
\title{Get a new predictions for output vertices after reweighting edges}
\usage{
getPrediction(g, v, new_weights)
}
\arguments{
\item{g,}{fitted graph}

\item{v,}{the vertex whose incoming edges will be updated}

\item{weights,}{a numeric vector containing the weights for those incoming edges.}
}
\description{
A given vertex has a given set of incoming edges, each of these edges has weights.
Upon supplying the vertex and weights on its incoming edges, this function updates the prediction
on the output edges.
}

