% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{gradientDescent}
\alias{gradientDescent}
\title{Gradient descent on weights}
\usage{
gradientDescent(wts_init, grad, loss, maxit = 100, epsilon = 0.01)
}
\arguments{
\item{loss}{function that takes a weight vector as an input and returns a single value}

\item{maxit}{maximum number of iterations}

\item{epsilon}{stop when change in loss output is less than epsilon 10 iterations in a row.}

\item{initial}{weight values}

\item{gradient}{function that takes weight vector as an input are returns a gradient vector of the same length}

\item{step}{step size}
}
\description{
Gradient descent on weights
}
\seealso{
\code{\link{gradientDescent}}, \code{\link{getGradientFunction}}, \code{\link{doChainRule}}
}

