% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signalgraph.R
\docType{data}
\name{historic_fits}
\alias{historic_fits}
\title{Fits of signal graph model to KEGG data}
\format{A list of lists, each list element containing the elements "fit" and "notes".}
\usage{
data(historic_fits)
}
\description{
A collection of signal graph fits to various graphs constructed from signaling pathways in KEGG.
}
\keyword{datasets}

