% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structure.R
\name{initializeEdges}
\alias{initializeEdges}
\title{Initalize edge attributes}
\usage{
initializeEdges(g)
}
\arguments{
\item{g}{igraph object. The vertices must be named.}
}
\value{
g igraph object
}
\description{
Used in contruction of a signalgraph.  Names and initializes the edges.
}

