% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structure.R
\name{initializeGraph}
\alias{initializeGraph}
\title{Priming an igraph bbject for fitting a signalgraph model}
\usage{
initializeGraph(g, data, graph_attr)
}
\arguments{
\item{g}{igraph object. The vertices must be named.}

\item{data}{a data frame. All of the names in the data from must match a vertex name.}

\item{graph_attr}{list of graph attributes.  Graph attributes include:
\itemize{
 \item{penalty}{penalized least squares error penalty parameter value}
 \item{activation}{the activation function (this actually is an R function), defaults to logistic.}
 \item{activation.prime}{The derivative fo the activation function, used in gradient calculation. Defaults to NULL}
 \item{min.max.constraints}{2 element numeric containing the acceptable range for each rate.}
 }}
}
\value{
A graph with all the attributes needed to fit the neural network model.
}
\description{
Any vertex that does not map to a name in the
}

