% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structure.R
\name{initializeVectorAttributesForVertex}
\alias{initializeVectorAttributesForVertex}
\title{Initialize numerically-valued vertex attributes}
\usage{
initializeVectorAttributesForVertex(g, v.index)
}
\arguments{
\item{g}{graph model}

\item{v.index}{vertex index}
}
\value{
and igraph object where the above attributes are initialized to the value \code{list(rep(NA, g$n))}
}
\description{
For each vertex, there are several sets of vectors representing data and calculations
made on the data.  These are stored as vertex attributes in the form of numeric objects
wrapped in a list so the length is one.  The vector attributes are;
\itemize{
 \item{input.signal}{The vector of linear combination of values from the parent nodes}
 \item{output.signal}{The output of the activation function applied to input.signal}
 \item{f.prime.input}{The derivative of the activation function applied to the input.signal}
 \item{observed}{Observed values in the data.  Not present for hidden and input nodes.}
 }
}
\details{
All of these are initialized with NA values
}

