% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structure.R
\name{initializeVertexBooleans}
\alias{initializeVertexBooleans}
\title{Initialize boolean vertex attibutes}
\usage{
initializeVertexBooleans(g)
}
\description{
There are 4 core boolean vertex attributes:
\itemize{
 \item{is.bias}{TRUE if the vertex is a bias or intercept}
 \item{is.observed}{TRUE if the vertex is observed in the data}
 \item{is.root}{TRUE if the vertex is a root in the graph}
 \item{is.leaf}{TRUE if the vertex is a leaf in the graph}
 }
}
\details{
This function initializes all these vertex attributes for later updating.
}

