% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structure.R
\name{initializeVertexVectors}
\alias{initializeVertexVectors}
\title{Add vector attributes to graph vertices}
\usage{
initializeVertexVectors(g)
}
\arguments{
\item{g}{igraph object}
}
\description{
signalgraph uses several vertex attributes that store a list of values.  These are
called vector attributes, because the set of values is used as a vector in calculations.
}

