% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structure.R
\name{initializeVertices}
\alias{initializeVertices}
\title{Initialize the vertex attributes and values}
\usage{
initializeVertices(g)
}
\description{
Initialize the vertex attributes and values
}
\seealso{
initializeVertexBooleans, addInterceptNodes, applyVertexVectors, addDataToVertices, resetUpdateAttributes
}

