% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structure.R
\name{initializeWeights}
\alias{initializeWeights}
\title{Simulating starting weights}
\usage{
initializeWeights(g)
}
\arguments{
\item{g}{igraph object}
}
\value{
a graph with updated weights
}
\description{
Starting weights used when fitting signalgraph objects are simulated from a uniform distribution
where the range is given by the 'min.max.constraints' attribute.
}

