% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{isValid}
\alias{isValid}
\title{Checks the validity of a numeric vertex attribute}
\usage{
isValid(item)
}
\description{
Checks the validity of a numeric vertex attribute
}
\examples{
g <- get_gate()
isValid(unlist(V(g)[type == "output"]$observed))
}

