% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structure.R
\name{isValidV}
\alias{isValidV}
\title{Checks the validity of a numeric vertex attribute}
\usage{
isValidV(item)
}
\description{
Checks the validity of a numeric vertex attribute
}
\examples{
g <- get_gate()
isValidV(unlist(V(g)[is.observed]$observed))
}

