% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structure.R
\name{loadCN}
\alias{loadCN}
\title{Load Markov Blankets}
\usage{
loadCN(g)
}
\arguments{
\item{g}{a signalgraph}
}
\description{
A Markov neighborhood for a given feature is the set of feature that includes that feature itself, and
all the nodes with which it shares a causal relationship.  If causality is given by a DAG, then it is
equivilent to the node and its Markov blanket.  This function adds this latter Markov blanket + identity 
based Markov neighborhood.
}

