% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{logistic}
\alias{logistic}
\alias{logistic_prime}
\title{The logistic function
The logistic function and its derivative (used in gradient calculations).}
\usage{
logistic(z)

logistic_prime(z)
}
\description{
The logistic function
The logistic function and its derivative (used in gradient calculations).
}

