% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{logistic_to_positive}
\alias{logistic_to_positive}
\title{Convert logistic activation parameters to parameters in u / (1 + u) function}
\usage{
logistic_to_positive(g, v)
}
\description{
Convert logistic activation parameters to parameters in u / (1 + u) function
}

