% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{long_test}
\alias{long_test}
\title{Perform long tests
 
Select whether or not to perform more time consuming tests by setting opt to TRUE or FALSE}
\usage{
long_test(option)
}
\description{
Perform long tests
 
Select whether or not to perform more time consuming tests by setting opt to TRUE or FALSE
}

