% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalgraph.R
\docType{data}
\name{mapk_g}
\alias{mapk_g}
\title{KEGG MAPK Signaling Pathway Phosphorylation Network}
\format{An igraph object with 60 vertices and 76 edges.  Vertex attributes include "label"
("MAP4K4", "PLA2G4B", "STMN1"...) and "kid" for KEGG ids ("hsa:9448", "hsa:100137049", "hsa:3925",...)}
\usage{
data(mapk_g)
}
\description{
A network of all the phosphorylation edges pulled from the MAPK Signaling Pathway on KEGG.
Edges from the human map ("hsa") are pulled.
\url{http://www.kegg.jp/kegg-bin/show_pathway?map=hsa04010&show_description=show}
}
\details{
The script for generating the graph is available here:
\url{https://gist.github.com/robertness/ecb79071301924f43d09}
}
\keyword{datasets}

