\name{neuralgraph-package}
\alias{neuralgraph-package}
\alias{neuralgraph}
\docType{package}
\title{
\packageTitle{neuralgraph}
}
\description{
\packageDescription{neuralgraph}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{neuralgraph}
\packageIndices{neuralgraph}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{neuralgraph}

Maintainer: \packageMaintainer{neuralgraph}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
