% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{newDataUpdate}
\alias{newDataUpdate}
\title{Add new data to fitted model in order to do prediction}
\usage{
newDataUpdate(g, input.table)
}
\description{
Add new data to fitted model in order to do prediction
}

