% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{order_edges}
\alias{order_edges}
\title{Order edges for ideal weighting
Recursively backtracks through structure and orders edges by proximity to the
target node, starting with the leaves.}
\usage{
order_edges(g, targets = get_leaves(g))
}
\value{
a numeric array of edge indices
}
\description{
Order edges for ideal weighting
Recursively backtracks through structure and orders edges by proximity to the
target node, starting with the leaves.
}

