% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{power_signal_graph}
\alias{power_signal_graph}
\title{Simulate a scale-free signal graph structure from an input structure}
\usage{
power_signal_graph(g, p, start_graph = NULL)
}
\arguments{
\item{g}{an igraph object.}

\item{p}{number of desired vertices in the output graph.}

\item{Starter}{graph as input for preferencial attachment.  Should have reversed edges. See start.graph argument in ba.game/barabasi.game/sample_pa functions in igraph.}
}
\value{
A new graph simulated from a power law based on g.
}
\description{
Takes an input graph structure, fits a power law model, and simulates a new
graph based on the fit.  This function uses the Barabasi-Albert model to 
generate graphs implemented in lucy::power_law_sim, however an alteration is made.  
The edges of the input graph are reversed before estimating the power law fit, and the 
simulated graph edges are reversed again before it is returned from the function.  
This way, the graph is characterized not by preferential attachment of incoming edges, but
by preferential attachment of outgoing edges.  Further, vertices with 0 degrees are avoided 
by having high degree vertices 'donate' edges to degreeless vertices.
}

