% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templates.R
\name{power_sim_no_singleton}
\alias{power_sim_no_singleton}
\title{Singleton-free power law sim
The objective is to simulate graphs with the same out-degree distribution as an input graph using
the Barabasi-Albert algorithm.  However, if there are nodes in the input graph that have no
out-degree, then there is a non-zero probability of simulating a it is possible the simulation
will produce a graph with singletons -- nodes with no edges.  To algorithm performs the
simulation normally, the removes the singletons from the graph. Using the resulting graph as a
starting graph, it continues simulating new nodes and edges but now with 0 probability of a node
having no outgoing edges, until the desired amount of  nodes is reached.  The output graph will
generally not have the same out-degree distribution, but will be close.  The lower the probability
of having 0 outgoing edges, the closer they will be.}
\usage{
power_sim_no_singleton(g, p)
}
\arguments{
\item{g}{the input igraph object.}

\item{p}{the desired amount of vertices}
}
\value{
a simulated graph based on g.
}
\description{
Singleton-free power law sim
The objective is to simulate graphs with the same out-degree distribution as an input graph using
the Barabasi-Albert algorithm.  However, if there are nodes in the input graph that have no
out-degree, then there is a non-zero probability of simulating a it is possible the simulation
will produce a graph with singletons -- nodes with no edges.  To algorithm performs the
simulation normally, the removes the singletons from the graph. Using the resulting graph as a
starting graph, it continues simulating new nodes and edges but now with 0 probability of a node
having no outgoing edges, until the desired amount of  nodes is reached.  The output graph will
generally not have the same out-degree distribution, but will be close.  The lower the probability
of having 0 outgoing edges, the closer they will be.
}

