% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templates.R
\name{rand_case}
\alias{rand_case}
\title{Test Case: Random graph and matching data}
\usage{
rand_case(m, k, n = m * k + m)
}
\arguments{
\item{m}{the number of desired nodes}

\item{k}{the number of desired observed nodes}

\item{n}{the number of desired rows in the data}
}
\value{
A list of two elements, graph and data.
}
\description{
Generates an igraph object and a corresponding dataset.  Used to
randomly generate test cases
}

