% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templates.R
\name{random_unfit_sg}
\alias{random_unfit_sg}
\title{Generate a random unfit signalgraph object}
\usage{
random_unfit_sg(k, n)
}
\arguments{
\item{k}{number of desired vertices}

\item{n}{number of desired points in the data}
}
\value{
an initialized, but unfit (unoptimized) signal graph model
}
\description{
Uses the generateMultiConnectedDAG in the lucy \url{https://github.com/osazuwa223/lucy} package.
Since the vertices that must have data are the roots and the vertices, (everything in between can be hidden),
then a data set is simulated for only those nodes.
}
\seealso{
initializeGraph
}

