% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templates.R
\name{regenerate_graph_data}
\alias{regenerate_graph_data}
\title{Regenerate the data values in a signal graph}
\usage{
regenerate_graph_data(g, n)
}
\arguments{
\item{g}{a signal graph object}

\item{n}{desired number of rows in output dataframe}
}
\value{
a dataframe of data
}
\description{
Simulates new data for all the fixed variables and propagates those changes
through the rest of the system.
}

