% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structure.R
\name{resetUpdateAttributes}
\alias{resetUpdateAttributes}
\title{Reset Model Attributes}
\usage{
resetUpdateAttributes(g)
}
\arguments{
\item{g}{a model}
}
\value{
A model with 'updated' attribute for each vertex is reset.
}
\description{
Resets the 'updated' vertex attribute for each vertex. All vertices, except for roots (which includes biases)
have this attribute set to FALSE.  The biases have this attribute set to TRUE.
}

