% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{sg_viz}
\alias{sg_viz}
\title{Visualize a signal graph object}
\usage{
sg_viz(g, main = NULL, sub = NULL, show_biases = FALSE)
}
\description{
Fixed nodes are displayed as orange.  Of the two kinds of random nodes,
observed nodes are displayed in green, hidden nodes are displayed in blue.
For random observed nodes, the width of node borders corresponds to error.
}

