\name{signalgraph-package}
\alias{signalgraph-package}
\alias{signalgraph}
\docType{package}
\title{
\packageTitle{signalgraph}
}
\description{
\packageDescription{signalgraph}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{signalgraph}
\packageIndices{signalgraph}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{signalgraph}

Maintainer: \packageMaintainer{signalgraph}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
