% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{sim_data_from_system}
\alias{sim_data_from_system}
\title{Generate data from a simulated signal graph
Pull data frame from a simulated signal graph and add error to simulate realistic data.}
\usage{
sim_data_from_system(g, factor)
}
\arguments{
\item{g}{the output of sim_system}

\item{factor}{numeric quantifying the amount of error to add to each random element in the 
data frame (see factor argument in base function 'jitter').}
}
\value{
a data frame that can be used as data from the simulated system.
}
\description{
Generate data from a simulated signal graph
Pull data frame from a simulated signal graph and add error to simulate realistic data.
}

