% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templates.R
\name{sim_system}
\alias{sim_system}
\title{Generate a random signalgraph object for data simulation}
\usage{
sim_system(m, n, error_sd = 0.2, ...)
}
\arguments{
\item{m}{the number of desired nodes}

\item{n}{the number of desired rows in the data}

\item{error_sd}{desired standard deviation for Gaussian error of logit(x)}

\item{...}{arguments past to fitNetwork, including graph attributes}
}
\value{
a signalgraph object
}
\description{
Produces a signalgraph objectwhere the values of the observed data and the fitted data are the same.
This is designed to produce a gold standard, which can simulate data, where upon a new model can be
fit on the simulated data, and the parameters of the standard and learned parameters of the new model
can be compared.
}

