% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templates.R
\name{sim_system_data}
\alias{sim_system_data}
\title{Simulate data from a signal graph}
\usage{
sim_system_data(g, n, add_error = FALSE)
}
\arguments{
\item{g}{a signal graph object}

\item{n}{desired number of rows in output dataframe}

\item{add_error}{if true, adds noise to each column in output dataframe using the jitter function.}
}
\value{
a dataframe of data
}
\description{
Simulates new data for all the fixed variables and propagates those changes
through the rest of the system.
}

