% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templates.R
\name{sim_system_from_graph}
\alias{sim_system_from_graph}
\title{Generate a random signalgraph object for data simulation based on an input graph structure}
\usage{
sim_system_from_graph(g, p, n, error_sd = 0.2, ...)
}
\arguments{
\item{g}{an igraph object, the output model graph structure is simulated from a power law estimated
from g.}

\item{p}{the number of desired vertices for the output model graph structure}

\item{n}{the number of desired data points for the output graph structure}

\item{error_sd}{desired amount of error}
}
\value{
a signal graph object
}
\description{
Produces a signalgraph object based on an input graph structure, such as a phosphorylation
graph.  A power law is fit to the input graph, and a new structure is simulated from that power
law.  Note that the input graph should be a good fit to a power law.
This is designed to produce a gold standard, which can simulate data, enabling you to fit a model
on the simulated data.  The parameters of the standard and learned parameters of the new model
can be compared.
}

