% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalgraph.R
\docType{data}
\name{titanic3}
\alias{titanic3}
\title{Titanic Data}
\format{A data frame with 14 variables and 1309 rows}
\source{
\url{http://biostat.mc.vanderbilt.edu/DataSets}
}
\usage{
data(titanic3)
}
\description{
Records of the survival of Titanic Passengers and such information as sex, age, fare each person paid, 
number of parents/children aboard, number of siblings or spouses aboard, passenger class and other fields.
}
\details{
Variable descriptions and how the data set was constructed: \url{http://biostat.mc.vanderbilt.edu/wiki/pub/Main/DataSets/titanic3info.txt}
}
\keyword{datasets}

