% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signalgraph.R
\docType{data}
\name{transformations}
\alias{transformations}
\title{Transformations}
\format{A data frame with 12 variables and 5 rows}
\usage{
data(transformations)
}
\description{
A artificial dataset for experimental purposes.  Contains two uniform random variables and
various transformations on those variables.  Some of the transformations have error added.
}
\details{
\itemize{
 \item A, simulated with runif(5)
 \item B, simulated with runif(5)
 \item A_squared = A^2
 \item A_root = A ^ .5,
 \item A_log = log(A),
 \item A_inverse = A ^ -1,
 \item AB_linear = .2 * A + .5 * B,
 \item AB_lin_err = AB_linear + Z,
 \item AB_rescaled = (AB_linear) /(1 + (AB_linear)),
 \item AB_scale1_nl_error =  (AB_lin_err) /(1 + (AB_lin_err)),
 \item AB_logistic = exp(AB_linear) /(1 + exp(AB_linear)),
 \item AB_logistic_err = exp(AB_lin_err) /(1 + exp(AB_lin_err))
}
}
\keyword{datasets}

