% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{update_signals}
\alias{update_signals}
\title{Update the 'signal' at each vertex in a signalgraph object}
\usage{
update_signals(g)
}
\arguments{
\item{g}{a signalgraph object.}
}
\value{
a signal graph object with updated values for the signal attributes.
}
\description{
Updates the 'output.signal' attribute of each vertex in the signalgraph.  Relies on the 'updateVertices'
propagation function in the lucy package, using calculateVals as the callback.
}

