% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{update_weights}
\alias{get_downstream_vertices}
\alias{order_vertices}
\alias{update_weights}
\title{Traversing the network for fit weight sets.}
\usage{
update_weights(g, verbose = FALSE)

order_vertices(g)

get_downstream_vertices(g, v)
}
\arguments{
\item{g}{a signalgraph object.}

\item{verbose}{if true prints messages about fitting performance.}
}
\value{
a signalgraph object with updated weights.
}
\description{
To reduce dimensionality of the problem of fitting all the weights in the graph, for each vertex
the weights on the incoming edges of that vertex are fit seperately.  Priority goes to 'fitted' vertices
(vertices that are used to calculate the mean squared error loss function), then vertices that
are upstream by 1 degree, then 2 degrees, etc.  'update_weights' traverses the vertices and
fits them individually. It uses 'vertex_updater" from the lucy package to apply the 'fit_weights_for_node'
callback on a vertex.  'vertex_updater' will first check all the downstream vertices for those that
are not yet updated (via the 'get_downstream_vertices' callback) and recursively fit those if they
are not yet updated.  However, 'order_vertices' is first applied to minimize the need for this
recursion.
}

