% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{updateSignals}
\alias{updateSignals}
\title{Update the 'signal' at each vertex in a signal graph}
\usage{
updateSignals(g)
}
\arguments{
\item{g}{a signalgraph object.}
}
\value{
g a signal graph object.
}
\description{
Updates the 'output.signal' attribute of each vertex in the signalgraph.  Relies on the 'updateVertices'
propagation function in the lucy package, using calculateVals as the callback.
}
\seealso{
calculateVals
}

