% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{vertexMSE}
\alias{vertexMSE}
\alias{vertexMSEs}
\title{Get the mean squared error for a given vertex}
\usage{
vertexMSE(g, v)

vertexMSEs(g)
}
\arguments{
\item{g}{a signalgraph object}

\item{v}{a vertex id in the signal graph}
}
\value{
means squared error calculation
}
\description{
vertexMSEs retrieves Get the mean squared error for all the random vertices
}

