% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{calculateVals}
\alias{calculateVals}
\title{Calculate the signal propagation in the signalgraph model}
\usage{
calculateVals(g, v)
}
\arguments{
\item{g,}{the graph model}

\item{v.index,}{the index of a vertex}
}
\value{
an updated graph model
}
\description{
For a given vertex, the values for the vertex attribute 'input.signal' are calculated as the linear
combination of outputs with edge weights as weights. 'output.signal' is calculated by applying the activation
function to the input signal.
}

