% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{fit_initialized_sg}
\alias{fit_initialized_sg}
\title{Fit an initialized network}
\usage{
fit_initialized_sg(g, epsilon = 1e-04, max.iter = 3)
}
\arguments{
\item{g}{an igraph object initialized to an unfit signalgraph object}

\item{epsilon}{when means square area falls below epsilon, stop}

\item{max.iter}{maximum number of iterations}
}
\description{
Fit an initialized network
}

