% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{fitInitializedNetwork}
\alias{fitInitializedNetwork}
\title{Fit an initialized network}
\usage{
fitInitializedNetwork(g, epsilon = 0.001, max.iter = 100, verbose = F)
}
\arguments{
\item{g}{an igraph object initialized to an unfit signalgraph object}

\item{epsilon}{when means square area falls below epsilon, stop}

\item{max.iter}{maximum number of iterations}

\item{verbose}{if TRUE print messages generated during optimization}
}
\description{
Fit an initialized network
}

