% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{fitNetwork}
\alias{fitNetwork}
\title{Fit a Neural Network on a Graph Structure}
\usage{
fitNetwork(g, input.table, output.table, epsilon = 0.05, max.iter = 3,
  activation = logistic, activation.prime = logistic_prime,
  min.max.constraints = NULL, verbose = F)
}
\description{
Fit a Neural Network on a Graph Structure
}

