% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{fitNetwork}
\alias{fitNetwork}
\title{Fit a Signalgraph model}
\usage{
fitNetwork(g, data, graph_attr = NULL, epsilon = 0.001, max.iter = 3)
}
\arguments{
\item{g}{igraph object. The vertices must be named.}

\item{data}{a data frame. All of the names in the data from must match a vertex name.}

\item{graph_attr}{list of graph attributes.  Graph attributes include:
\itemize{
 \item{L1_pen}{penalized least squares error L1 penalty parameter value}
 \item{L2_pen}{penalized least squares error L2 penalty parameter value}
 \item{activation}{the activation function (this actually is an R function), defaults to logistic.}
 \item{activation.prime}{The derivative fo the activation function, used in gradient calculation. Defaults to NULL}
 \item{min.max.constraints}{2 element numeric containing the acceptable range for each rate.}
 }}

\item{epsilon}{when means square area falls below epsilon, stop}

\item{max.iter}{maximum number of iterations}
}
\value{
A fitted signalgraph object.
}
\description{
Initializes a graph into a signalgraph then fits the model.
}

