% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{get_fitted}
\alias{get_fitted}
\title{Get a data frame of output signals}
\usage{
get_fitted(g)
}
\arguments{
\item{g}{A fitted graph}
}
\value{
a data frame containing fitted values
}
\description{
Get a data frame of output signals
}
\seealso{
recover_design
}

