% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{getDependentEdges}
\alias{getDependentEdges}
\title{Find edges that affect an edge's optimization}
\usage{
getDependentEdges(g, e)
}
\arguments{
\item{g}{a signalgraph object}

\item{e}{an edge index in the signal graph}
}
\value{
set of edge indices
}
\description{
When optimizing weights by back-propagation, before optimizing the objective function for a
given weight, it is necessary to first optimize for all weights that are no paths downstream
from the given weight and upstream of an observed variable.  This function identifies all such
downstream edges relative an input edge.
}

