% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backpropagation.R
\name{getMSE}
\alias{getMSE}
\title{MSE calculation over random variables in the graph}
\usage{
getMSE(g)
}
\arguments{
\item{g}{a signal graph object}
}
\description{
Calculates the mean squared error for all the random variables in a signalgraph
object, then reports the mean of means.
}

