% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimization.R
\name{getObjective}
\alias{getObjective}
\title{Closure for generating a penalized least squares loss function for use in optimization}
\usage{
getObjective(initial_graph, v)
}
\arguments{
\item{initial_graph}{an initialized signal graph, the prefix 'initial' in 'initial_graph' is in contrast to
candidate graphs against which the initial_graph will be compared in the optimization.}

\item{v}{the vertex whose weights will be optimized.}
}
\description{
This function is a closure that creates a loss function that takes a weight vector as an argument.
The closure takes a graph vertex as an argument.  The algorithm identifives the incoming edges to the vertex,
then pulls the weights from those edges, constructing a weight vector.  This weight vector is the argument
for the loss function, and is used as the starting values in the optimization. The function calculates squared
error loss by first taking this weight vector input and predicting the outcome variable, then calculating loss
from the predicted and observed values.
}

