% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{getOptimizationFunction}
\alias{getOptimizationFunction}
\title{Closure for optimization with BFGS and specified gradient function
Called by fit_weights_for_node.  Creates a closure based on the 'optim' function.  Uses
BFGS or unbounded BFGS depending on if bounds are provided.}
\usage{
getOptimizationFunction(g, lossFunction, getGradient)
}
\description{
Closure for optimization with BFGS and specified gradient function
Called by fit_weights_for_node.  Creates a closure based on the 'optim' function.  Uses
BFGS or unbounded BFGS depending on if bounds are provided.
}
\seealso{
\code{\link{fit_weights_for_node}}, \code{\link{getObjective}}, \code{\link{getOptimizationFunction}}, \code{\link{getOptimizationFunctionNG}}, \code{\link{getPrediction}}
}

