% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backpropagation.R
\name{getOptimizationFunctionNG}
\alias{getOptimizationFunctionNG}
\title{CLosure for optimization with BFGS, no gradient function specified}
\usage{
getOptimizationFunctionNG(g, lossFunction)
}
\description{
CLosure for optimization with BFGS, no gradient function specified
}

