% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{getPrediction}
\alias{getPrediction}
\title{Get a new graph used for prediction}
\usage{
getPrediction(g, v, new_weights)
}
\arguments{
\item{g,}{fitted graph}

\item{v,}{the vertex whose incoming edges will be updated}

\item{weights,}{a numeric vector containing the weights for those incoming edges.}
}
\value{
a new updated graph that can be used to generate a prediction.
}
\description{
A given vertex has a given set of incoming edges, each of these edges has weights.
Upon supplying the vertex and weights on its incoming edges, this function updates the graph and
returns the updated graph object.
}
\seealso{
\code{\link{fit_weights_for_node}}, \code{\link{getObjective}}, \code{\link{getOptimizationFunction}}, \code{\link{getOptimizationFunctionNG}}, \code{\link{getPrediction}}
}

