% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structure.R
\name{initializeGraphAttributes}
\alias{initializeGraphAttributes}
\title{Initialize graph attributes of a graph object}
\usage{
initializeGraphAttributes(g, graph_attr, n)
}
\arguments{
\item{g}{an igraph object not yet initialized as a signal graph}

\item{graph_attr}{a list of objects to be used as graph attributes}

\item{n}{number of rows in the data}
}
\description{
Signalgraph stores key parameters used in fitting and other operations on the signal graph as igraph
attributes.  igraph has 3 kinds of attributes, edge attributes, vertex attributes, and graph attributes.
This function sets graph attributes.  Key examples include:
 \itemize{
 \item{L1_pen}{penalized least squares error L1 penalty parameter value}
 \item{L2_pen}{penalized least squares error L2 penalty parameter value}
 \item{activation}{the activation function (this actually is an R function)}
 \item{activation.prime}{The derivative fo the activation function, used in gradient calculation.}
 \item{min.max.constraints}{2 element numeric containing the acceptable range for each rate.}
 }
}

