% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{initializeVertexVectors}
\alias{initializeVertexVectors}
\title{Initialize numerically-valued vertex attributes}
\usage{
initializeVertexVectors(g, v.index)
}
\arguments{
\item{g}{graph model}

\item{v.index}{vertex index}
}
\value{
and igraph object where the above attributes are initialized to the value \code{list(rep(NA, g$n))}
}
\description{
For a given vertex, vertex attibutes that take a numeric vector as a value are
initialized with a placeholder. Specifically, the vertex attributes are;
\itemize{
 \item{input.signal}{The vector of linear combination of values from the parent nodes}
 \item{output.signal}{The output of the activation function applied to input.signal}
 \item{f.prime.input}{The derivative of the activation function applied to the input.signal}
 \item{observed}{Observed values in the data.  Not present for hidden and input nodes.}
 }
All of these are initialized with NA values
}

