% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structure.R
\name{initializeVertices}
\alias{initializeVertices}
\title{Initialize the vertex attributes and values}
\usage{
initializeVertices(g, data)
}
\arguments{
\item{g}{an igraph object}
}
\description{
Initializes the boolean attributes of the vectors,
 \enumerate{
  \item Initialize the boolean vertex attributes
  \item Add bias nodes nodes
  \item Add vector attributes to vertices
  \item Populate vector attributes with data
  \item Reset the 'update' vertex and edge attributes
}
}

