% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templates.R
\name{power_signal_graph}
\alias{power_signal_graph}
\title{Simulate a scale-free signal graph structure based on another}
\usage{
power_signal_graph(g, n)
}
\arguments{
\item{g}{an igraph object.}

\item{n}{number of desired vertices in the output graph.}
}
\value{
A new graph simulated from a power law based on g.
}
\description{
Takes an input signal graph, fits a power law model, and simulates a new
graph based on the fit.  This function uses the Barabasi-Albert model to
generate graphs implemented in lucy::power_signal_graph, however an alteration is made.
The edges of the input graph are reversed before estimating the power law fit, and the
simulated graph edges are reversed again before it is returned from the function.
This way, the graph is characterized not by preferential attachment of incoming edges, but
by preferential attachment of outgoing edges.
}

