% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templates.R
\name{random_fit_sg}
\alias{random_fit_sg}
\title{Generate a random fitted signalgraph object}
\usage{
random_fit_sg(m, k, n, max.iter = 1, ...)
}
\arguments{
\item{m}{the number of desired nodes}

\item{k}{the number of desired observed nodes}

\item{n}{the number of desired rows in the data}

\item{...}{additional arguments, including graph attributes}
}
\value{
a signalgraph object
}
\description{
Uses the generateMultiConnectedDAG in the lucy \url{https://github.com/osazuwa223/lucy} package.
The vertices that must have data are the roots and the leaves, (everything in between can be hidden),
so data is simulated for only those nodes.
}

