% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{random_sg}
\alias{random_sg}
\title{Generate a random fitted signalgraph object}
\usage{
random_sg(p, n, max.iter = 1, input_g = NULL, method = "ordered",
  no_fixed = FALSE, ...)
}
\arguments{
\item{p}{the number of desired nodes}

\item{n}{the number of desired rows in the data}

\item{input_g}{an igraph object.  If supplied then the graph is simulated from
a power law fit on this input graph.}

\item{method}{the fitting method for simulating a directed acyclic graph.  Ignored
if input_g is supplied.}

\item{no_fixed}{boolean if TRUE then all vertices are treated as random. Defaults to FALSE}

\item{...}{additional arguments, including graph attributes}
}
\value{
a signalgraph object
}
\description{
Uses the sim_DAG in the lucy \url{https://github.com/robertness/lucy} package.  
The vertices that must have data are the roots and the leaves, (everything in between can be hidden),
so data is simulated for only those nodes.
}

