% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{recover_design}
\alias{recover_design}
\title{Pull observed data for the inputs and outputs of a signal graph model into a data frame}
\usage{
recover_design(g)
}
\arguments{
\item{g}{A fitted graph}
}
\value{
a data frame, essentially the design matrix.
}
\description{
Pull the observed data (AKA examples) for the input and output vertices of a signal graph model into a data frame.
}
\seealso{
get_fitted
}

