% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{recover_design}
\alias{recover_design}
\title{Pull observed data from a signalgraph object}
\usage{
recover_design(g)
}
\arguments{
\item{g}{A fitted graph}
}
\value{
a data frame, essentially the design matrix.
}
\description{
Pull the observed data (AKA examples) for the observed vertices in a signal graph model into a data frame.
}
\seealso{
get_fitted
}

