% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structure.R
\name{resetUpdateAttributes}
\alias{resetUpdateAttributes}
\title{Reset Model Attributes}
\usage{
resetUpdateAttributes(g)
}
\arguments{
\item{g}{a igraph object}
}
\value{
A model with 'updated' attribute for each vertex and edge is reset.
}
\description{
Resets the 'updated' attribute for each vertex and edge. All vertices, except for roots (which includes biases)
have this attribute set to FALSE.  The biases have this attribute set to TRUE -- this attribute should always
be true for biases.
}

