% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{sg_viz}
\alias{sg_viz}
\title{Visualize a signal graph object}
\usage{
sg_viz(g, max_val = 1, main = NULL, sub = NULL, show_biases = FALSE)
}
\arguments{
\item{max_val}{A normalizing denominator used in calculating thickness of red error
border.  Should be at least as large as the highest expected variation.  If there are
vertices with MSE values greater than max_val a warning is thrown.  Defaults to 1/12
the MSE expected for a random variable between 0, 1 with maximum entropy (i.e a uniform on 0, 1).}
}
\description{
Fixed nodes are displayed as orange.  Of the two kinds of random nodes, 
observed nodes are displayed in green, hidden nodes are displayed in blue.
For random observed nodes, the width of node borders corresponds to error.
}

