% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{update_signals}
\alias{calculate_signal}
\alias{getLinearCombination}
\alias{update_signals}
\title{Propagation of fitted values through the graph.}
\usage{
update_signals(g, verbose = FALSE)

calculate_signal(g, v)

getLinearCombination(wts, model.mat)
}
\arguments{
\item{g}{a signalgraph object.}

\item{v}{the vertex to be updated given the parents}

\item{wts}{the weights of the incoming edges given the parents}

\item{model.mat}{a matrix constructing of the "output.signal" attributes of the parents}
}
\value{
a signal graph object with updated values for the signal attributes.
}
\description{
'update_signals' uses graph propagation to update the fitted values (the output.signal attribute) at each vertex
 in a signalgraph object.  It relies on the graph propagation function 'update_vertices' in the lucy
 package, using calculate_signal as the callback.
 The 'calculate_signal' callback, when applied to a given vertex, calculates the linear combination
 of fitted values from the parent nodes, and then applies the activation function to calculate 
 the 'output.signal' attribute.  The linear combination is calculated in the function 'getLinearCombination'.
}

