% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{update_weights}
\alias{update_weights}
\title{Update the weights for the incoming edges of a vertex in a signalgraph object}
\usage{
update_weights(g, verbose = FALSE)
}
\arguments{
\item{g}{a signalgraph object.}
}
\value{
a signalgraph object with updated weights.
}
\description{
Updates the 'weight' attribute of each incoming edge for a vertex in the signalgraph.
Relies on the 'update_vertices' propagation function in the lucy package,
using calculate_signal as the callback.
}

