% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{update_weights}
\alias{update_weights}
\title{Update the weight at each edge in a signalgraph object}
\usage{
update_weights(g)
}
\arguments{
\item{g}{a signalgraph object.}
}
\value{
a signalgraph object with updated weights.
}
\description{
Updates the 'weight' attribute of each edge in the signalgraph.  Relies on the 'updateedges'
propagation function in the lucy package, using calculateVals as the callback.
}

